fixed timeline duplication for multi-day shoots
timeline name duplication will cause the later one will not be created at all.
simply add a shooting day in front of the timeline names to distinguish them.
solve this problem will also solve the problem of `get_timeline_by_name` function,
because now the names of all timelines in the entire project are unique.